% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{best_model_description}
\alias{best_model_description}
\title{Provide information about the best model}
\usage{
best_model_description(models_list)
}
\arguments{
\item{models_list}{The list of fitted models}
}
\value{
Nothing but display the results.
}
\description{
This is a main function of this package. It displays the information about
the best model.
}
\examples{
best_model_description(models_list = Models)

}
