% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{boot_all_models}
\alias{boot_all_models}
\title{Refit all models several times on bootstrapped data}
\usage{
boot_all_models(models_list, data, n_boot = 40L, xtol_rel = 1e-07,
  maxtime = Inf, n_CPU = 2L, save_rda = FALSE, seed = 1L)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{data}{The dataset to be used}

\item{n_boot}{The number of bootstrapped dataset to generate and fit}

\item{xtol_rel}{Tolerance parameter for nloptr}

\item{maxtime}{Maximum time (in seconds) allowed for nloptr to fit; use Inf for best fit!}

\item{n_CPU}{The number of CPU to use for fitting the bootstrapped datasets}

\item{save_rda}{A boolean indicating whether saving the model as a *.rda object or not}

\item{seed}{An integer giving the seed for the (pseudo-)random generator}
}
\value{
The list containing all the fits.
}
\description{
This is a main function of this package. It refits each model multiple times
after performing a non-parametric bootstrap on the data. This is used to
ultimately derive confidence intervals on parameter estimates.
}
\examples{
## Warning: increase n_boot and maxtime for real fits!
Models_boot_test <- boot_all_models(models_list = Models[1:2],
                                    data = SurvEles_small,
                                    n_boot = 2L, maxtime = 1)
Models_boot_test

}
\seealso{
\code{\link{fit_all_models}}
}
