% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Build the design matrix}
\usage{
build_design_matrix(data, param_matrix, indices = NULL,
  empty_param_matrix_full = NULL, empty_design_matrix = NULL,
  print = TRUE)
}
\arguments{
\item{data}{The dataset to be used}

\item{param_matrix}{The matrix storing parameter values}

\item{indices}{The list of indices computed by \code{compute_indices}}

\item{empty_param_matrix_full}{The empty expanded or complete parameter matrix}

\item{empty_design_matrix}{The empty design matrix}

\item{print}{A boolean indicating whether or not to print information about potential issues}
}
\value{
The design matrix.
}
\description{
This function is called internally by other functions.
It builds the design matrix for the models.
}
\examples{
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
dmat <- build_design_matrix(data = SurvEles_small, param_matrix = pmat)
table(dmat[, "w1"])

}
