% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_aic_w}
\alias{compute_aic_w}
\title{Compute the AIC weights}
\usage{
compute_aic_w(models_list)
}
\arguments{
\item{models_list}{The list of fitted models}
}
\value{
The vector of AIC weights.
}
\description{
This function is called internally by other functions.
It computes the AIC weights from a list of fitted models.
}
\examples{
compute_aic_w(Models)

}
\seealso{
\code{\link{table_boot}}
}
