% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{compute_demography}
\alias{compute_demography}
\title{Provide demographic information about the data}
\usage{
compute_demography(data)
}
\arguments{
\item{data}{The dataset to be used}
}
\value{
Nothing, only displays the outcome.
}
\description{
This is a main function of this package. It displays the information about
the demography of elephants that we reported in the paper.
}
\examples{
compute_demography(data = SurvEles_small)

}
