% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_estimates_average}
\alias{compute_estimates_average}
\title{Compute parameter estimates by model averaging}
\usage{
compute_estimates_average(models_list)
}
\arguments{
\item{models_list}{The list of fitted models}
}
\value{
A vector of estimates.
}
\description{
This function is called internally by other functions.
It computes parameter estimates by model averaging.
}
\details{
We used the so-called "shrinkage estimator".
That is, the parameter estimates from all models are being considered
and parameters that were not estimated are considered to be equal to zero.
}
\examples{
compute_estimates_average(models_list = Models)

}
\seealso{
\code{\link{table_model_averaging}}
}
