% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_indices}
\alias{compute_indices}
\title{Compute the indices}
\usage{
compute_indices(data)
}
\arguments{
\item{data}{The dataset to be used}
}
\value{
A names list of indices.
}
\description{
This function is called internally by other functions.
It computes indices (i.e. row numbers) referring to different characteristics of the elephants which are required
to build the design matrix of the mode.
}
\examples{
indices <- compute_indices(data = SurvEles_small)
lapply(indices, head)  ## display first 6 indices of each element

}
\seealso{
\code{\link{build_design_matrix}}
}
