% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{compute_logLik}
\alias{compute_logLik}
\title{Compute the log-likelihood}
\usage{
compute_logLik(pred_prob, surv_bin, scale = 1L)
}
\arguments{
\item{pred_prob}{The predicted probability of death computed by \code{predictor()}}

\item{surv_bin}{The binary response variable taking 1 for alive and 0 for death}

\item{scale}{The scaling parameter (usefull if the optimisation function can only minimise)}
}
\value{
The scaled log-likelihood.
}
\description{
This function is not actually used. It is a R replicate of the C++ function \code{compute_logLikC}
that is actually used. It computes the log-likelihood.
}
\examples{
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
dmat <- build_design_matrix(data = SurvEles_small, param_matrix = pmat)
p <- predictor(data = SurvEles_small, design_matrix = dmat)
compute_logLik(pred_prob = p, surv_bin = SurvEles_small$Alive, scale = 1L)
compute_logLikC(pred_prob = p, surv_bin = SurvEles_small$Alive, scale = 1L)

}
