% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_logLikC}
\alias{compute_logLikC}
\title{C++ Function computing the log-likelihood}
\usage{
compute_logLikC(pred_prob, surv_bin, scale)
}
\arguments{
\item{pred_prob}{The predicted probability of death computed by \code{predictor()}}

\item{surv_bin}{The binary response variable taking 1 for alive and 0 for death}

\item{scale}{The scaling parameter (useful if the optimisation function can only minimise)}
}
\description{
This function is a C++ implementation of \code{\link{compute_logLik}}.
It is called internally by other functions.
}
\seealso{
\code{\link{compute_logLik}}
}
