% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{compute_survival_info}
\alias{compute_survival_info}
\title{Provide survival information}
\usage{
compute_survival_info(models_list)
}
\arguments{
\item{models_list}{The list of fitted models}
}
\value{
Nothing, only displays the outcome.
}
\description{
This is a main function of this package. It displays the information about
the survival of elephants that we repported in the paper (e.g. half life...).
}
\examples{
compute_survival_info(models_list = Models)

}
