% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ElesAll}
\alias{ElesAll}
\title{The dataset of information at the individual level for all elephants}
\format{A \var{dataframe} with 5150 elephants.}
\description{
This dataset contains the breakdowns for the counts of elephants used in the
paper.
}
\examples{

## Basic information about the data:
head(ElesAll)
str(ElesAll)
sum(ElesAll$N)

## Number of elephants in each sex:
tapply(ElesAll$N, ElesAll$Sex, sum)

## Number of elephants in each region:
tapply(ElesAll$N, ElesAll$Region, sum)

}
\references{
This study.
}
\seealso{
SurvEles_small, ElesCaptured
}
\keyword{datasets}
