% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ElesCaptured}
\alias{ElesCaptured}
\title{Datasets of information at the individual level for captured elephant}
\format{A \var{list} with two \var{data.frame} describing the 2072 captured elephants.}
\description{
This object contains a list with two datasets containing the breakdowns for
the counts of captured elephants used in the paper. One list provides
breakdowns per capture age and the other one provides breakdowns per decade.
}
\examples{

## Basic information about the data:
lapply(ElesCaptured, head)
lapply(ElesCaptured, function(i) sum(i$N))

## Number of elephants by age at capture:
ElesCaptured$ElesCapturedByAge

## Number of elephants by decade at capture:
ElesCaptured$ElesCapturedByDecade

}
\references{
This study.
}
\seealso{
SurvEles_small, ElesAll
}
\keyword{datasets}
