% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{evidence_ratios_top_models}
\alias{evidence_ratios_top_models}
\title{Build the table comparing the top models}
\usage{
evidence_ratios_top_models(models_list)
}
\arguments{
\item{models_list}{The list of fitted models}
}
\value{
A data frame containing the results.
}
\description{
This function provides evidence ratios and AIC differences among top models.
The top models are the set of best models for which the sum of weights amount
to 95\% of the sum on all 17 models.
}
\examples{
evidence_ratios_top_models(models_list = Models)

}
