% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_param_matrix}
\alias{expand_param_matrix}
\title{Expand the parameter matrix}
\usage{
expand_param_matrix(param_matrix, param_matrix_full = NULL)
}
\arguments{
\item{param_matrix}{The matrix storing parameter values}

\item{param_matrix_full}{The expanded or complete parameter matrix}
}
\value{
The expanded parameter matrix.
}
\description{
This function is called internally by other functions.
Since the parameter matrix created with \code{build_param_matrix} only
contains the parameter values for the parameter that needs to be optimised,
the parameter matrix varies in length. This causes troubles to work with the
other functions, so this function expands the matrix to its full size by
replicating parameter values for the different parameters of the same meta
parameter which have not been defined when building the parameter matrix (see
example).
}
\examples{
pmat <- build_param_matrix(param_values = seq(0.01, 0.21, length = 21),
                           param_str = c("w1" = "1", "w3" = "1+t+l", "w4" = "s", "w5" = "c"))
pmat
expand_param_matrix(pmat)

}
