% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{Fit the model}
\usage{
fit(data, param_matrix_ini, xtol_rel = 1e-07, maxtime = Inf, lb = 0,
  ub = 3, print_level = 0)
}
\arguments{
\item{data}{The dataset to be used}

\item{param_matrix_ini}{The matrix storing initial values for the parameters}

\item{xtol_rel}{Tolerance parameter for nloptr}

\item{maxtime}{Maximum time (in seconds) allowed for nloptr to fit; use Inf for best fit!}

\item{lb}{Lower boundary for parameter values during the optimisation (default = 0)}

\item{ub}{Upper boundary for parameter values during the optimisation (default = 3)}

\item{print_level}{Printing setting for nloptr}
}
\value{
A list containing the outcome of the fitting procedure
}
\description{
This function is called internally by \code{\link{fit_all_models}}.
It is the actual function performing the fitting procedure.
To fit all the models, use \code{fit_all_models} instead.
}
\details{
The function uses the computation of the likelihood from \code{\link{logLik_from_vector_param}} and
optimise the parameter values using the optimisation routine \code{\link[nloptr]{nloptr}} from the package of the
same name. Specifically, we are using the method called "NLOPT_LN_BOBYQA".
}
\examples{
pmat_ini <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
pmat_ini
## Warning: increase maxtime for real fit!
fit(data = SurvEles_small, param_matrix_ini = pmat_ini, maxtime = 1)

}
\seealso{
\code{\link{fit_all_models}}
}
