% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_all_models}
\alias{fit_all_models}
\title{Fit all 17 models}
\usage{
fit_all_models(data, xtol_rel = 1e-07, maxtime = Inf, n_CPU = 2L,
  save_rda = FALSE)
}
\arguments{
\item{data}{The dataset to be used}

\item{xtol_rel}{Tolerance parameter for nloptr}

\item{maxtime}{Maximum time (in seconds) allowed for nloptr to fit; use Inf for best fit!}

\item{n_CPU}{The number of CPU to use for fitting the bootstrapped datasets}

\item{save_rda}{A boolean indicating whether saving the model as a *.rda object or not}
}
\value{
The list of fitted models.
}
\description{
This is a main function of this package. It fits all main 17 models presented in the paper.
}
\examples{
## Warning: increase maxtime for real fit!
models_test <- fit_all_models(data = SurvEles_small, maxtime = 1)
models_test

}
\seealso{
\code{\link{fit}}
}
