% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_param_name}
\alias{get_param_name}
\title{This function selects the names of parameters belonging to the same meta parameter}
\usage{
get_param_name(meta_param_name = NULL, param_names)
}
\arguments{
\item{meta_param_name}{The name of the meta parameter}

\item{param_names}{A vector of parameter names}
}
\value{
A vector of parameter names.
}
\description{
This function is called internally by other functions. We distinguish meta
parameters (e.g. "w1") from actual parameters (e.g. "w1.females",
"w1.males"). This function allows for the retrieval of all parameters that
have the same meta parameter name. If no name is given for the meta
parameter, the function extracts the meta parameter names from the parameter
names.
}
\examples{
get_param_name(meta_param_name = "w1", param_names = rownames(Models$mod1$param_matrix_full))

}
\seealso{
\code{\link{get_param_names_without_meta}}
}
