% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{interpol_surv}
\alias{interpol_surv}
\title{Interpolate yearly survivorship}
\usage{
interpol_surv(death_prob, ages, steps = 10)
}
\arguments{
\item{death_prob}{A vector of yearly probabilities of death}

\item{ages}{A vector of ages for which to perform the interpolation}

\item{steps}{An integer giving the number of increments for the interpolation (within each year)}
}
\value{
A data frame containing the interpolated survivorship probabilities for each age.
}
\description{
This function is called internally by other functions. It interpolates yearly
survivorship probabilities. The method is described in Slud 2001.
}
\examples{
interpol_surv(death_prob = c(0.1, 0.8, 0.9, 0.9), ages = 1:4)

}
\references{
Slud 2001 Actuarial Mathematics and Life-Table Statistics
(\url{http://www2.math.umd.edu/~hmg/Stat470book.pdf})
}
