% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{list_tables_boot}
\alias{list_tables_boot}
\title{Build the list of tables showing the results of the non-parametric bootstrap on each model}
\usage{
list_tables_boot(models_list, models_list_boot, print = TRUE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{print}{A boolean indicating whether or not to print information about potential issues}
}
\value{
A list storing the tables created by \code{\link{table_boot}}.
}
\description{
This function is called internally by other functions.
It processes the outcomes of the fit of each of the 17 models on different bootstrap
replicates.
}
\details{
The table(s) produced contain the following columns:
\enumerate{
  \item \code{beta} the parameter estimate in the original fit
  \item \code{mean} the mean of the parameter estimates across the fits on the bootstrapped data
  \item \code{bias} the difference between the column \var{mean} and \var{beta}
  \item \code{SE} the standard _error_ of the parameter estimates computed as the standard _deviation_ of the parameter estimates across the fits on the bootstrapped data
  \item \code{lwr} the lower boundary of the 95\% confidence interval approximated as \code{pnorm(0.025)*SE}
  \item \code{upr} the upper boundary of the 95\% confidence interval approximated as \code{pnorm(0.975)*SE}
}
}
\examples{
list_tables_boot(models_list = Models, models_list_boot = Models_boot)

}
\seealso{
\code{\link{boot_all_models}}, \code{\link{table_boot}}, \code{\link{table_model_averaging}}
}
