% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{logLik_from_design_matrix}
\alias{logLik_from_design_matrix}
\title{Compute the log-likelihood from the design matrix}
\usage{
logLik_from_design_matrix(data, design_matrix, scale = 1, indices = NULL)
}
\arguments{
\item{data}{The dataset to be used}

\item{design_matrix}{The design matrix}

\item{scale}{The scaling parameter (usefull if the optimisation function can only minimise)}

\item{indices}{The list of indices computed by \code{compute_indices}}
}
\value{
The scaled log-likelihood.
}
\description{
This function is called internally by other functions.
It prepares data and call \code{\link{predictorC_wrap}} to predict the probability of death
and then call the C++ function \code{compute_logLikC} which computes the log-likelihood in C++.
}
\examples{
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
dmat <- build_design_matrix(data = SurvEles_small, param_matrix = pmat)
logLik_from_design_matrix(data = SurvEles_small, design_matrix = dmat)

}
\seealso{
\code{\link{predictorC_wrap}}, \code{\link{compute_logLik}}
}
