% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{logLik_from_vector_param}
\alias{logLik_from_vector_param}
\title{Compute the log-likelihood from parameter vector}
\usage{
logLik_from_vector_param(param_values, param_str, data, indices = NULL,
  scale = 1, param_matrix = NULL, empty_param_matrix_full = NULL,
  empty_design_matrix = NULL)
}
\arguments{
\item{param_values}{The vector of values for the parameters}

\item{param_str}{The named vector indicating the parameter structure, that is, the type of subcategory for each meta parameter}

\item{data}{The dataset to be used}

\item{indices}{The list of indices computed by \code{compute_indices}}

\item{scale}{The scaling parameter (usefull if the optimisation function can only minimise)}

\item{param_matrix}{The matrix storing parameter values}

\item{empty_param_matrix_full}{The empty expanded or complete parameter matrix}

\item{empty_design_matrix}{The empty design matrix}
}
\value{
The scaled log-likelihood.
}
\description{
This function is called internally by other functions.
It is a wrapper to the function \code{\link{logLik_from_design_matrix}} which has
inputs consistent the the requirements of optimisation routines.
}
\details{
If the argument \code{param_matrix} is given, the parameter matrix will be
updated and not build from scratch. This will save some computation time.
}
\examples{
logLik_from_vector_param(param_values = c(0.1, 0.200001), param_str = c("w1" = "s"),
                         data = SurvEles_small)
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
logLik_from_vector_param(param_values = c(0.1, 0.200001, rep(0, 14)), param_str = c("w1" = "s"),
                         data = SurvEles_small, param_matrix = pmat)

}
\seealso{
\code{\link{logLik_from_design_matrix}}
}
