% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_param_str_list}
\alias{make_param_str_list}
\title{Create the list of parameter structures}
\usage{
make_param_str_list()
}
\value{
The list storing the parameter structure for all 17 models.
}
\description{
This function is called internally by other functions. It creates the
attributes \code{param_str} for each of the 17 models to be fitted.
}
\examples{
make_param_str_list()

}
