% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{median_lifespan}
\alias{median_lifespan}
\title{Compute the median lifespan}
\usage{
median_lifespan(newdata, models_list)
}
\arguments{
\item{newdata}{A data frame providing the covariate values for predictions}

\item{models_list}{The list of fitted models}
}
\value{
The value of the median lifespan.
}
\description{
This function is called internally by other functions. It computes the
predicted median lifespan.
}
\examples{
d <- data.frame(Sex = "males", CaptureMethod = "CAPTIVE",
                CaptureAge = 0, Age = 0:100, TimeSinceCapture = 0,
                Region = "Mandaley", BirthCohort = "1960")
median_lifespan(newdata = d, models_list = Models)

}
