% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Models}
\alias{Models}
\title{The list of the fitted models}
\format{A \var{list} with the 17 outputs of the function \code{fit_all_models}.}
\description{
This dataset contains the 17 models used in the paper. These models have been
fitted on the complete dataset!
}
\examples{
## The names of the models:
names(Models)

## The AIC of model 1:
Models[["mod1"]]$AIC

## The convergence status of model 1:
Models[["mod1"]]$message

## Studying the effect of potential convergence issues:
table(unlist(lapply(Models, function(mod) mod$convergence))) ## 1 = perfect convergence
### Conclusion: there is no sign of convergence issue.
### All 17 models fitted without any trouble.

## The estimates from model1:
data.frame(estimates = Models[["mod1"]]$param_matrix[, "value"])

}
\references{
This study.
}
\keyword{datasets}
