% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Models_boot}
\alias{Models_boot}
\title{The list of the models refitted on bootstrapped data}
\format{A \var{list} with the 680 outputs of the function \code{CI_all_models}.}
\description{
This dataset contains the 40 bootstraps replicate for each of the 17 models.
These models have been fitted on the complete dataset!
}
\examples{
## The names of the models:
names(Models_boot)

## The AIC of model 1 replicate 1:
Models_boot[["mod1_1"]]$AIC

## The convergence status of model 1 replicate 1:
Models_boot[["mod1_1"]]$message

## The estimates from model 1 replicate 1:
data.frame(estimates = Models_boot[["mod1_1"]]$param_matrix[, "value"])

## Studying the effect of potential convergence issues:
table(unlist(lapply(Models_boot, function(mod) mod$convergence))) ## 1 = perfect convergence
models_with_pb <- which(unlist(lapply(Models_boot, function(mod) mod$convergence))!=1)
CI_with_pb <- table_model_averaging(models_list = Models,
                                    models_list_boot = Models_boot)
CI_without_pb <- table_model_averaging(models_list = Models,
 models_list_boot = Models_boot[!(names(Models_boot) \%in\% names(models_with_pb))])
max(abs(100*((CI_with_pb$SE - CI_without_pb$SE) /  CI_without_pb$SE)), na.rm = TRUE)
### Conclusion: only 3 fits out of 680 reached maximal time allowed before
### reaching full convergence. Those 3 model fits impact the SE of parameter
### estimates by less than 1.8\% in the worst situation, which is negligeable.
### We thus chose not to exclude them and base all CI computation on all 680
### fits.


}
\references{
This study.
}
\keyword{datasets}
