% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_function_parameters.R
\name{parameters}
\alias{parameters}
\title{Definition for parameters}
\usage{
parameters(age_at_capture, ages, binary, capture_method, check, data,
  death_prob, design_matrix, empty_design_matrix, empty_param_matrix_full,
  file_name, fit_original, indices, lb, maxtime, meta_param_name,
  meta_param_type, model, models_list, models_list_boot, n_boot, n_CPU,
  name_first_column, newdata, param_matrix, param_matrix_full,
  param_matrix_ini, param_names, param_str, param_values, pred_prob, print,
  print_level, save_pdf, save_rda, save_xlsx, scale, seed, sex, steps,
  surv_bin, table, ub, xtol_rel)
}
\arguments{
\item{age_at_capture}{The age at capture}

\item{ages}{A vector of ages for which to perform the interpolation}

\item{binary}{A boolean indicated if the binary version of the package must be built}

\item{capture_method}{A character indicating the method of capture of the elephant: \var{"CAPTIVE"}, \var{"IMM"}, \var{"MILARSHI"} or \var{"STOCKADE"}}

\item{check}{Whether the arguments should be checked against expected types}

\item{data}{The dataset to be used}

\item{death_prob}{A vector of yearly probabilities of death}

\item{design_matrix}{The design matrix}

\item{empty_design_matrix}{The empty design matrix}

\item{empty_param_matrix_full}{The empty expanded or complete parameter matrix}

\item{file_name}{The file name for the exported table}

\item{fit_original}{A fitted model}

\item{indices}{The list of indices computed by \code{compute_indices}}

\item{lb}{Lower boundary for parameter values during the optimisation (default = 0)}

\item{maxtime}{Maximum time (in seconds) allowed for nloptr to fit; use Inf for best fit!}

\item{meta_param_name}{The name of the meta parameter}

\item{meta_param_type}{The type of subcategories for the meta parameters}

\item{model}{A fitted model}

\item{models_list}{The list of fitted models}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{n_boot}{The number of bootstrapped dataset to generate and fit}

\item{n_CPU}{The number of CPU to use for fitting the bootstrapped datasets}

\item{name_first_column}{The name of the first column which is taken from the row names of the table}

\item{newdata}{A data frame providing the covariate values for predictions}

\item{param_matrix}{The matrix storing parameter values}

\item{param_matrix_full}{The expanded or complete parameter matrix}

\item{param_matrix_ini}{The matrix storing initial values for the parameters}

\item{param_names}{A vector of parameter names}

\item{param_str}{The named vector indicating the parameter structure, that is, the type of subcategory for each meta parameter}

\item{param_values}{The vector of values for the parameters}

\item{pred_prob}{The predicted probability of death computed by \code{predictor()}}

\item{print}{A boolean indicating whether or not to print information about potential issues}

\item{print_level}{Printing setting for nloptr}

\item{save_pdf}{A boolean indicating whether saving the plot as a *.pdf object or not}

\item{save_rda}{A boolean indicating whether saving the model as a *.rda object or not}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}

\item{scale}{The scaling parameter (useful if the optimisation function can only minimise)}

\item{seed}{An integer giving the seed for the (pseudo-)random generator}

\item{sex}{A character indicating the sex of the elephant: \var{"males"} or \var{"females"}}

\item{steps}{An integer giving the number of increments for the interpolation (within each year)}

\item{surv_bin}{The binary response variable taking 1 for alive and 0 for death}

\item{table}{The table to be exported as *.xlsx}

\item{ub}{Upper boundary for parameter values during the optimisation (default = 3)}

\item{xtol_rel}{Tolerance parameter for nloptr}
}
\description{
Here is the list of all parameters used in the package.
}
