% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_baseline}
\alias{plot_baseline}
\title{Plot the baseline mortality rate}
\usage{
plot_baseline(data, models_list, save_pdf = FALSE)
}
\arguments{
\item{data}{The dataset to be used}

\item{models_list}{The list of fitted models}

\item{save_pdf}{A boolean indicating whether saving the plot as a *.pdf object or not}
}
\value{
Nothing, the function only plots.
}
\description{
This is a main function of this package. It plots the baseline mortality rate.
}
\examples{
plot_baseline(data = SurvEles_small, models_list = Models)

}
