% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_captured_demog}
\alias{plot_captured_demog}
\title{Plot demographic information for the captured elephants}
\usage{
plot_captured_demog(data, save_pdf = FALSE)
}
\arguments{
\item{data}{The dataset to be used}

\item{save_pdf}{A boolean indicating whether saving the plot as a *.pdf object or not}
}
\value{
Nothing, the function only plots.
}
\description{
This is a main function of this package. It plots information about when
elephants have been captured.
}
\examples{
plot_captured_demog(ElesCaptured)

}
\seealso{
ElesCaptured
}
