% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_survivorship}
\alias{plot_survivorship}
\title{Plot the survivorship curves}
\usage{
plot_survivorship(data, models_list, save_pdf = FALSE)
}
\arguments{
\item{data}{The dataset to be used}

\item{models_list}{The list of fitted models}

\item{save_pdf}{A boolean indicating whether saving the plot as a *.pdf object or not}
}
\value{
Nothing, the function only plots.
}
\description{
This is a main function of this package. It plots the survivorship curve for
each capture method.
}
\details{
The survivorship curves are averages of individual predictions matching each
observation. They thus capture the effects of differences in sample size
according to birth cohorts, regions, time since capture... between
different capture methods.
}
\examples{
plot_survivorship(data = SurvEles_small, models_list = Models)

}
