% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict}
\alias{predict}
\title{Predict mortality based on a single model}
\usage{
predict(model, newdata)
}
\arguments{
\item{model}{A fitted model}

\item{newdata}{A data frame providing the covariate values for predictions}
}
\value{
A vector with the mortality prediction(s).
}
\description{
This function is called internally by other functions. It predicts the
mortality rate (per year) of elephants based on a covariate information
provided as a data frame and based on one model.
}
\examples{
d <- data.frame(Sex = "males", CaptureMethod = "STOCKAGE",
                CaptureAge = 20, Age = 30, TimeSinceCapture = 10,
                Region = "Kachin", BirthCohort = "2000")
predict(model = Models[[1]], newdata = d)

}
