% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_avg}
\alias{predict_avg}
\title{Predict mortality using model averaging}
\usage{
predict_avg(models_list, newdata)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{newdata}{A data frame providing the covariate values for predictions}
}
\value{
A vector with the mortality prediction(s).
}
\description{
This function is called internally by other functions. It predicts the
mortality rate (per year) of elephants based on a covariate information
provided as a data frame and based on a list of models used to compute the
parameter estimates by model averaging.
}
\examples{
d <- data.frame(Sex = "males", CaptureMethod = "STOCKAGE",
                CaptureAge = 20, Age = 30, TimeSinceCapture = 10,
                Region = "Kachin", BirthCohort = "2000")
predict_avg(models_list = Models, newdata = d)

}
