% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{predictor}
\alias{predictor}
\title{Predict the probability of death}
\usage{
predictor(data, design_matrix, indices = NULL)
}
\arguments{
\item{data}{The dataset to be used}

\item{design_matrix}{The design matrix}

\item{indices}{The list of indices computed by \code{compute_indices}}
}
\value{
A vector of the probability of death.
}
\description{
This function is not actually used. It is a R replicate of the C++ function \code{predictorC}
that is actually used. It computes the probability of death for each observation.
}
\examples{
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
dmat <- build_design_matrix(data = SurvEles_small, param_matrix = pmat)
p <- predictor(data = SurvEles_small, design_matrix = dmat)
table(p)

}
\seealso{
\code{\link{predictorC}}
}
