% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictorC}
\alias{predictorC}
\title{C++ Function computing the predicted probabilities}
\usage{
predictorC(w1, b1, w2, b2, w3, w4, w5, b4, b5, age, timecap,
  indices_captured)
}
\arguments{
\item{w1}{The vector of the corresponding name extracted from the design matrix}

\item{b1}{The vector of the corresponding name extracted from the design matrix}

\item{w2}{The vector of the corresponding name extracted from the design matrix}

\item{b2}{The vector of the corresponding name extracted from the design matrix}

\item{w3}{The vector of the corresponding name extracted from the design matrix}

\item{w4}{The vector of the corresponding name extracted from the design matrix}

\item{w5}{The vector of the corresponding name extracted from the design matrix}

\item{b4}{The vector of the corresponding name extracted from the design matrix}

\item{b5}{The vector of the corresponding name extracted from the design matrix}

\item{age}{The vector of the corresponding name extracted from the data}

\item{timecap}{The vector of the corresponding name extracted from the data}

\item{indices_captured}{The vector of boolean indicated captured individuals}
}
\description{
This function is a C++ implementation of the predictor function.
It is called internally by other functions.
}
\seealso{
\code{\link{predictor}}
}
