% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{predictorC_wrap}
\alias{predictorC_wrap}
\title{Predict the probability of death (R wrapper for C++)}
\usage{
predictorC_wrap(data, design_matrix, indices = NULL)
}
\arguments{
\item{data}{The dataset to be used}

\item{design_matrix}{The design matrix}

\item{indices}{The list of indices computed by \code{compute_indices}}
}
\value{
A vector of the probability of death.
}
\description{
This function is called internally by other functions.
It is a wrapper for the function \code{predictorC} which
computes the probability of death in C++ (for faster computation).
}
\examples{
pmat <- build_param_matrix(param_values = c(0.1, 0.2), param_str = c("w1" = "s"))
dmat <- build_design_matrix(data = SurvEles_small, param_matrix = pmat)
p <- predictorC_wrap(data = SurvEles_small, design_matrix = dmat)
table(p)

}
\seealso{
\code{\link{predictor}}
}
