% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{prepare_data_for_capture_plot}
\alias{prepare_data_for_capture_plot}
\title{Prepare the data for plotting the effect of capture on the mortality rate}
\usage{
prepare_data_for_capture_plot(models_list, data, age_at_capture = 0, sex,
  capture_method)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{data}{The dataset to be used}

\item{age_at_capture}{The age at capture}

\item{sex}{A character indicating the sex of the elephant: \var{"males"} or \var{"females"}}

\item{capture_method}{A character indicating the method of capture of the elephant: \var{"CAPTIVE"}, \var{"IMM"}, \var{"MILARSHI"} or \var{"STOCKADE"}}
}
\value{
A list containing the data required for the plotting function.
}
\description{
This function is called internally by other functions. It prepares the data
for the function \code{\link{plot_capture}}.
}
\examples{
res <- prepare_data_for_capture_plot(models_list = Models,
                                     data = SurvEles_small,
                                     age_at_capture = 5,
                                     sex = "males",
                                     capture_method = "IMM")
lapply(res, range)

}
\seealso{
\code{\link{plot_capture}}
}
