% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{save_xlsx}
\alias{save_xlsx}
\title{Save a table in the *.xlsx format}
\usage{
save_xlsx(table, file_name, name_first_column = NULL)
}
\arguments{
\item{table}{The table to be exported as *.xlsx}

\item{file_name}{The file name for the exported table}

\item{name_first_column}{The name of the first column which is taken from the row names of the table}
}
\value{
The input table (invisibly).
}
\description{
This function is called internally by other functions. It is a wrapper for
the function \code{\link[openxlsx:write.xlsx]{write.xlsx}} of the package \pkg{openxlsx}.
}
