% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SurvEles_small}
\alias{SurvEles_small}
\title{A subset of the dataset for elephant survival}
\format{A \var{dataframe} with 100 elephant year observations.}
\description{
This dataset contains 100 rows of the 83026 ones used in the real analysis.
Unfortunately, we cannot share the original data due to property restriction.
The subset can be used to test all our functions. It has the same structure
as the complete dataset we used (see example below).
}
\note{
See \code{\link{ElesAll}} for the processed dataset containing
information on all individuals from the complete dataset.
}
\examples{

## Basic information about the data:
head(SurvEles_small)
str(SurvEles_small)
nrow(SurvEles_small)

## Number of observations in each sex:
table(SurvEles_small$Sex)

}
\references{
This study.
}
\seealso{
ElesAll, ElesCaptured
}
\keyword{datasets}
