% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_all_models}
\alias{table_all_models}
\title{Build the table comparing all 17 models}
\usage{
table_all_models(models_list, save_xlsx = FALSE, print = TRUE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}

\item{print}{A boolean indicating whether or not to print information about potential issues}
}
\value{
The table of comparing all model fits.
}
\description{
This is a main function of this package. It creates the table comparing all
17 models contained in the paper.
}
\examples{
table_all_models(Models)


}
\seealso{
\code{\link{fit_all_models}}
}
