% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_boot}
\alias{table_boot}
\title{Build the table of the results of the non-parametric bootstrap}
\usage{
table_boot(model, models_list_boot, print = TRUE)
}
\arguments{
\item{model}{A fitted model}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{print}{A boolean indicating whether or not to print information about potential issues}
}
\value{
A data frame storing the results.
}
\description{
This function is called internally by other functions.
It processes the outcomes of the fit of the same model on different bootstrap
replicates.
}
\details{
The table(s) produced contain the following columns:
\enumerate{
  \item \code{beta} the parameter estimate in the original fit
  \item \code{mean} the mean of the parameter estimates across the fits on the bootstrapped data
  \item \code{bias} the difference between the column \var{mean} and \var{beta}
  \item \code{SE} the standard _error_ of the parameter estimates computed as the standard _deviation_ of the parameter estimates across the fits on the bootstrapped data
  \item \code{lwr} the lower boundary of the 95\% confidence interval approximated as \code{pnorm(0.025)*SE}
  \item \code{upr} the upper boundary of the 95\% confidence interval approximated as \code{pnorm(0.975)*SE}
}
}
\examples{
index <- unlist(lapply(strsplit(names(Models_boot), "_"),
                function(i) i[1] == "mod1")) ## find which bootstrap belong to mod1
mod1_boot <- Models_boot[index]
table_boot(model = Models$mod1, models_list_boot = mod1_boot)

}
\seealso{
\code{\link{boot_all_models}}, \code{\link{list_tables_boot}}, \code{\link{table_model_averaging}}
}
