% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_estimates}
\alias{table_estimates}
\title{Build the table with model parameter estimates for all models}
\usage{
table_estimates(models_list, models_list_boot, save_xlsx = FALSE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}
}
\value{
A data.frame containing the estimates.
}
\description{
This is a main function of this package. It builds the table containing all
parameter estimates for all models, as well as those produce produced by
model averaging.
}
\examples{
table_estimates(models_list = Models,
                models_list_boot = Models_boot)

}
\seealso{
\code{\link{fit_all_models}}, \code{\link{boot_all_models}}
}
