% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_lifespan}
\alias{table_lifespan}
\title{Build the table with the predicted median lifespans.}
\usage{
table_lifespan(models_list, save_xlsx = FALSE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}
}
\value{
A data frame containing the median lifespans.
}
\description{
This is a main function of this package. It builds the table containing the
prediction for the median lifespans for the different categories of elephants.
The predictions are done for two ages at capture (5 and 20 years old).
}
\examples{
table_lifespan(models_list = Models)

}
