% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_model_averaging}
\alias{table_model_averaging}
\title{Build the table of the results of the model averaging across the 17 fitted models}
\usage{
table_model_averaging(models_list, models_list_boot, save_xlsx = FALSE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}
}
\value{
A data.frame storing the output of the model average.
}
\description{
This is a main function of this package. It performs the model averaging.
}
\details{
To compute averaged parameter estimates, we used the so-called "shrinkage estimator".
That is, the parameter estimates from all models are being considered
and parameters that were not estimated are considered to be equal to zero.

To compute the confidence intervals, we used the approximation [\code{pnorm(0.025)*SE}; \code{pnorm(0.975)*SE}].

To compute the standard error, we used equation 5 in Lukacs et al. 2010 (see References).

We indicated estimates higher than 1e-5 by a star stored in the last column of the table.
}
\examples{
table_model_averaging(models_list = Models,
                      models_list_boot = Models_boot)


}
\references{
Lukacs, Burnham & Anderson (2010). Model selection bias and Freedman’s paradox.
Ann. Inst. Stat. Math. 62:117–125
}
\seealso{
\code{\link{boot_all_models}}, \code{\link{table_boot}}, \code{\link{list_tables_boot}}, \code{\link{table_SE}}
}
