% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_SE}
\alias{table_SE}
\title{Build a table with the standard error of all parameter estimates for all models}
\usage{
table_SE(models_list, models_list_boot, print = TRUE, save_xlsx = FALSE)
}
\arguments{
\item{models_list}{The list of fitted models}

\item{models_list_boot}{The list of fitted models after bootstrap}

\item{print}{A boolean indicating whether or not to print information about potential issues}

\item{save_xlsx}{A boolean indicating whether saving the table as a *.xlsx document or not}
}
\value{
A data.frame storing the standard errors.
}
\description{
This function is called internally by other functions.
It is used for performing the model averaged computation of the confidence
intervals on parameter estimates. It provides the standard error on each parameter
estimates as estimated by bootstrap.
}
\examples{
table_SE(models_list = Models, models_list_boot = Models_boot, print = FALSE)

}
\seealso{
\code{\link{table_model_averaging}}
}
