% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{upload_on_drat}
\alias{upload_on_drat}
\title{Build and upload the package on drat}
\usage{
upload_on_drat(binary = FALSE)
}
\arguments{
\item{binary}{A boolean indicated if the binary version of the package must be built}
}
\description{
This is a function for development purpose only. Drat is a package that
allows for the installation of R packages from GitHub. We use this system to
distribute the package (see Readme.Rmd). The function defined here builds the
package, commits the changes and push them on the drat repository on GitHub.
It is designed to work on the maintainer computer only.
}
